+++ element_zstack/volume_matching.py
@@ -12,10 +12,7 @@ schema = dj.Schema()
 
 
 def activate(
-    schema_name: str,
-    *,
-    create_schema: bool = True,
-    create_tables: bool = True
+    schema_name: str, *, create_schema: bool = True, create_tables: bool = True
 ):
     """Activate this schema.
     Args:
@@ -37,11 +34,24 @@ def activate(
 
 @schema
 class VolumeMatchTask(dj.Manual):
+    """Defines the volumes and segmentations that will be registered in the downstream `VolumeMatch` table.
+
+    Attributes:
+        volume_match_task (uuid): UUID hash for the volume matching task
+    """
+
     definition = """  # 
     volume_match_task: uuid
     """
 
     class Volume(dj.Part):
+        """Defines the volume segmentations that will be registered in the downstream `VolumeMatch` table.
+
+        Attributes:
+            VolumeMatchTask (foreign key): Primary key from `VolumeMatchTask`.
+            volume.Segmentation (foreign key): Primary key from `volume.Segmentation`.
+        """
+
         definition = """
         -> master
         -> volume.Segmentation
@@ -49,18 +59,26 @@ class VolumeMatchTask(dj.Manual):
 
     @classmethod
     def insert1(cls, vol_seg_keys, **kwargs):
-        """
+        """Insert an entry into the `VolumeMatchTask` and `VolumeMatchTask.Volume` tables.
+
         Args:
             vol_seg_keys (tuple): a tuple of two cell-segmented volumes
         """
-        assert len(vol_seg_keys) == 2, f"Volume match task only supports matching two cell-segmented volumes, {len(vol_seg_keys)} are provided"
+        assert (
+            len(vol_seg_keys) == 2
+        ), f"Volume match task only supports matching two cell-segmented volumes, {len(vol_seg_keys)} are provided"
         vol_seg_keys = [(volume.Segmentation & k).fetch1("KEY") for k in vol_seg_keys]
-        assert len(set(vol_seg_keys)) == 2, "The two specified cell-segmented volumes are identical"
+        assert (
+            len(set(vol_seg_keys)) == 2
+        ), "The two specified cell-segmented volumes are identical"
 
         hashed = hashlib.md5()
-        [hashed.update(str(k).encode()) for k in sorted([dict_to_uuid(k) for k in vol_seg_keys])]
+        [
+            hashed.update(str(k).encode())
+            for k in sorted([dict_to_uuid(k) for k in vol_seg_keys])
+        ]
 
-        mkey = {'volume_match_task': uuid.UUID(hex=hashed.hexdigest())}
+        mkey = {"volume_match_task": uuid.UUID(hex=hashed.hexdigest())}
         if cls & mkey:
             assert len(cls.Volume & mkey & vol_seg_keys) == 2
             return
@@ -72,6 +90,14 @@ class VolumeMatchTask(dj.Manual):
 
 @schema
 class VolumeMatch(dj.Computed):
+    """Execute the volume matching algorithm and store the results.
+
+    Attributes:
+        VolumeMatchTask (foreign key): Primary key from `VolumeMatchTask`.
+        execution_time (datetime): Execution time of the volume matching task.
+        execution_duration (float): Duration of the volume matching task.
+    """
+
     definition = """
     -> VolumeMatchTask
     ---
@@ -80,6 +106,15 @@ class VolumeMatch(dj.Computed):
     """
 
     class Transformation(dj.Part):
+        """Store the transformation matrix to the common space.
+
+        Attributes:
+            VolumeMatch (foreign key): Primary key from `VolumeMatch`.
+            VolumeMatchTask.Volume (foreign key): Primary key from `VolumeMatchTask.Volume`.
+            transformation_matrix (longblob): the transformation matrix to
+            transform to the common space.
+        """
+
         definition = """  # transformation matrix
         -> master
         -> VolumeMatchTask.Volume
@@ -88,11 +123,29 @@ class VolumeMatch(dj.Computed):
         """
 
     class CommonMask(dj.Part):
+        """Store common mask identifier.
+
+        Attributes:
+            common_mask (smallint): Integer value for the common mask identifier.
+        """
+
         definition = """
         common_mask: smallint
         """
 
     class VolumeMask(dj.Part):
+        """For the masks in the common space, store the associated mask in the segmented volumes, and the confidence of the volume registration and cell matching.
+
+        Attributes:
+            VolumeMatch.CommonMask (foreign key): Primary key from
+            `VolumeMatch.CommonMask`.
+            VolumeMatchTask.Volume (foreign key): Primary key from
+            `VolumeMatchTask.Volume`.
+            volume.Segmentation.Mask (foreign key): Primary key from
+            `volume.Segmentation.Mask`.
+            confidence (float): confidence in the volume registration and cell matching between the segmented volumes.
+        """
+
         definition = """
         -> master.CommonMask
         -> VolumeMatchTask.Volume
@@ -105,10 +158,13 @@ class VolumeMatch(dj.Computed):
         import point_cloud_registration as pcr
         from scipy.stats import gaussian_kde
 
-        vol_keys = (volume.Segmentation & (VolumeMatchTask.Volume & key)).fetch('KEY')
+        vol_keys = (volume.Segmentation & (VolumeMatchTask.Volume & key)).fetch("KEY")
 
-        vol1_points, vol2_points = zip(*(volume.Segmentation.Mask & vol_keys).fetch(
-            'mask_center_x', 'mask_center_y', 'mask_center_z'))
+        vol1_points, vol2_points = zip(
+            *(volume.Segmentation.Mask & vol_keys).fetch(
+                "mask_center_x", "mask_center_y", "mask_center_z"
+            )
+        )
 
         vol1_points = np.hstack([*vol1_points])
         vol2_points = np.hstack([*vol2_points])