+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 Observes [Semantic Versioning](https://semver.org/spec/v2.0.0.html) standard and
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) convention.
 
-## [0.1.0] - Unreleased
+## [0.1.0] - 2023-05-10
 
-+ Add - `zstack` schema
++ Add - `volume` schema
++ Add - `bossdb` schema
++ Add - Upload utilty to create resources and upload data to BossDB
 
 [0.1.0]: https://github.com/datajoint/element-zstack/releases/tag/0.1.0